Move Test Behaviors CreatesRequest, HasFAker, and ReadsResponseContent out of Unit directory
