+++ test/HasFaker.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace RCatlin\Blog\Test\Unit;
+namespace RCatlin\Blog\Test;
 
 use Faker\Factory;
