+++ test/Unit/Controller/Api/ArticleCreateControllerTest.php
@@ -10,10 +10,10 @@ use RCatlin\Blog\Entity;
 use RCatlin\Blog\ReverseTransformer;
 use RCatlin\Blog\Serializer;
 use RCatlin\Blog\ServiceProvider;
+use RCatlin\Blog\Test\CreatesRequest;
+use RCatlin\Blog\Test\HasFaker;
+use RCatlin\Blog\Test\ReadsResponseContent;
 use RCatlin\Blog\Test\Unit\BuildsMocks;
-use RCatlin\Blog\Test\Unit\CreatesRequest;
-use RCatlin\Blog\Test\Unit\HasFaker;
-use RCatlin\Blog\Test\Unit\ReadsResponseContent;
 use RCatlin\Blog\Validator;
 use Refinery29\Piston\Request;
 use Refinery29\Piston\Response;
