+++ test/Unit/ReverseTransformer/Entity/TagReverseTransformerTest.php
@@ -5,8 +5,8 @@ namespace RCatlin\Blog\Test\Unit\ReverseTransformer\Entity;
 use RCatlin\Blog\Entity;
 use RCatlin\Blog\Repository;
 use RCatlin\Blog\ReverseTransformer;
+use RCatlin\Blog\Test\HasFaker;
 use RCatlin\Blog\Test\Unit\BuildsMocks;
-use RCatlin\Blog\Test\Unit\HasFaker;
 
 class TagReverseTransformerTest extends \PHPUnit_Framework_TestCase
 {
