+++ test/Unit/Serializer/TransformerContainerTest.php
@@ -4,8 +4,8 @@ namespace RCatlin\Blog\Test\Unit\Serializer;
 
 use League\Container\Container;
 use League\Fractal\TransformerAbstract;
+use RCatlin\Blog\Test\HasFaker;
 use RCatlin\Blog\Test\Unit\BuildsMocks;
-use RCatlin\Blog\Test\Unit\HasFaker;
 use RCatlin\Blog\Transformer;
 
 class TransformerContainerTest extends \PHPUnit_Framework_TestCase
