+++ test/Unit/Transformer/Entity/TagTransformerTest.php
@@ -3,7 +3,7 @@
 namespace RCatlin\Blog\Test\Unit\Transformer\Entity;
 
 use RCatlin\Blog\Entity;
-use RCatlin\Blog\Test\Unit\HasFaker;
+use RCatlin\Blog\Test\HasFaker;
 use RCatlin\Blog\Transformer\Entity\TagTransformer;
 
 class TagTransformerTest extends \PHPUnit_Framework_TestCase
